refactor(jetbrains): extract session scrolling
